# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Laravel API Swagger documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz/{quiz}/start": {
            "post": {
                "tags": [
                    "Quiz Attempts"
                ],
                "summary": "Start a quiz attempt",
                "operationId": "a5a5006c40ee1965d2867f353fb1ba32",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Ro'ziyev Sanjarbek"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz attempt started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attempt_id": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/attempts/answer": {
            "post": {
                "tags": [
                    "Quiz Attempts"
                ],
                "summary": "Submit an answer for a question",
                "operationId": "d8f994d4fc2f275c565a0a1c270218f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attempt_id",
                                    "question_id",
                                    "selected_option_id"
                                ],
                                "properties": {
                                    "attempt_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "question_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "selected_option_id": {
                                        "type": "integer",
                                        "example": 42
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "your_answer": {
                                            "type": "string",
                                            "example": "Laravel is a PHP framework"
                                        },
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "correct_answer": {
                                            "type": "string",
                                            "example": "Laravel is a PHP framework"
                                        },
                                        "correct_option_id": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quiz/finish": {
            "post": {
                "tags": [
                    "Quiz Attempts"
                ],
                "summary": "Finish quiz attempt and calculate results",
                "operationId": "88090d4a751820efb1b2c78201c388df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attempt_id"
                                ],
                                "properties": {
                                    "attempt_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "full_name": {
                                            "type": "string",
                                            "example": "Ro'ziyev Sanjarbek"
                                        },
                                        "correct": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "wrong": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "percentage": {
                                            "type": "number",
                                            "example": 70
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/attempt-answers": {
            "get": {
                "tags": [
                    "Quiz Attempts"
                ],
                "summary": "Get all attempts with answers",
                "operationId": "41ef67125ca5e0fbf33a09a1ac7b625d",
                "responses": {
                    "200": {
                        "description": "Attempts fetched successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}